% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_coords.R
\name{circle_coords}
\alias{circle_coords}
\title{Circle Coordinates}
\usage{
circle_coords(lat = lat, lon = lon, adj_size = adj_size)
}
\arguments{
\item{lat}{Latitude of the center of the circle. Default is 'lat'.}

\item{lon}{Longitude of the center of the circle. Default is 'lon'.}

\item{adj_size}{Adjustment size for the circle. This will affect the radius of the circle. Default is 'adj_size'.}
}
\value{
A Polygon object representing the circle.
}
\description{
This function generates the coordinates for a circle given a center (latitude and longitude) and an adjustment size.
The circle is defined using the parametric representation, x = rcos(theta) and y = rsin(theta).
}
\examples{
\dontrun{
circle_coords(lat = 40.7128, lon = -74.0060, adj_size = 1)
}
}
