% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_build.R
\name{dict_build}
\alias{dict_build}
\title{Construct the dictionary}
\usage{
dict_build()
}
\value{
A dataframe that serves as a dictionary for a classification scheme.
}
\description{
This function builds a dataframe that serves as a dictionary for a classification scheme.
The dictionary includes six fields: "group" in Portuguese and English, "class" in Portuguese
and English, a numerical "code", and a "color" associated with each class.
}
