% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_public_gdrive_file.R
\name{download_public_gdrive_file}
\alias{download_public_gdrive_file}
\title{Download Public Google Drive File}
\usage{
download_public_gdrive_file(file_link, local_path)
}
\arguments{
\item{file_link}{Link to the file on Google Drive.}

\item{local_path}{Path to save the file locally.}
}
\value{
No return value, called for side effects.
}
\description{
This function downloads a file from a public Google Drive link.
}
\examples{
\dontrun{
  download_public_gdrive_file(file_link = "https://drive.google.com/file/d/xxxx", local_path = "./local_dir/my_file.txt")
}
}
