% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gdrive_filename.R
\name{get_gdrive_filename}
\alias{get_gdrive_filename}
\title{Get Google Drive Filename}
\usage{
get_gdrive_filename(file_link)
}
\arguments{
\item{file_link}{The shareable link of the Google Drive file.}
}
\value{
The filename of the Google Drive file.

The filename of the Google Drive file.
}
\description{
This function retrieves the filename of a Google Drive file from its shareable link.
}
\examples{
\dontrun{
get_gdrive_filename("https://drive.google.com/file/d/1m9xxBPaDQz9xxBPaDQz9/view")
}
\dontrun{
get_gdrive_filename("https://drive.google.com/file/d/1m9xxBPaDQz9xxBPaDQz9/view")
}
}
