% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagon_coords.R
\name{hexagon_coords}
\alias{hexagon_coords}
\title{Hexagon Coordinates}
\usage{
hexagon_coords(lat = lat, lon = lon, adj_size = adj_size)
}
\arguments{
\item{lat}{Latitude of the center of the hexagon. Default is 'lat'.}

\item{lon}{Longitude of the center of the hexagon. Default is 'lon'.}

\item{adj_size}{Adjustment size for the hexagon. This will affect the size of the hexagon. Default is 'adj_size'.}
}
\value{
A Polygon object representing the hexagon.
}
\description{
This function generates the coordinates for a regular hexagon given a center (latitude and longitude) and an adjustment size.
The hexagon's vertices are calculated based on the geometric properties of a regular hexagon.
}
\examples{
\dontrun{
hexagon_coords(lat = 40.7128, lon = -74.0060, adj_size = 1)
}
}
