% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_dist.R
\name{land_dist}
\alias{land_dist}
\title{Land Distribution for a Specified Year}
\usage{
land_dist(data = NULL, year = NULL, type = "barplot")
}
\arguments{
\item{data}{A list that includes 'aggregate_data', a dataframe with the calculated areas for each land class and year.}

\item{year}{A specific year to visualize the land distribution.}

\item{type}{The type of plot to generate, either "pie" or "barplot".}
}
\value{
A plot showing the distribution of different land classes in a specific year.
}
\description{
This function creates a plot (either pie chart or barplot) showing the distribution of different land classes in a specific year.
}
\examples{
\dontrun{
land_dist(data = result, year = 1985, type = "barplot")
}
}
