% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_vis.R
\name{land_vis}
\alias{land_vis}
\title{Land Class Visualization}
\usage{
land_vis(data = NULL, year = 1985)
}
\arguments{
\item{data}{A list containing at least a raster list element with the raster images. If NULL, the function will stop and print an error message.}

\item{year}{A numeric value indicating the year to be visualized. Should be between 1985 and 2020. If out of these bounds, the function will stop and print an error message. Default is 1985.}
}
\value{
A ggplot object displaying the land classes for the specified year.
}
\description{
This function generates a map visualization for land classes in a specified year.
}
\examples{
\dontrun{
land_vis(data = my_data, year = 1990)
}

}
