% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rasters_from_folder.R
\name{load_rasters_from_folder}
\alias{load_rasters_from_folder}
\title{Load Rasters From Folder}
\usage{
load_rasters_from_folder(start, end, folder_path)
}
\arguments{
\item{start}{The start year for which raster files will be loaded.}

\item{end}{The end year for which raster files will be loaded.}

\item{folder_path}{The path to the folder containing the raster files.}
}
\value{
A list of raster files loaded from the specified folder path for each year in the range from start to end.
}
\description{
This function loads raster files from a specified folder path for a given range of years.
The raster files are expected to be named in the format "brasil_coverage_year.tif".
}
\examples{
\dontrun{
load_rasters_from_folder(2000, 2005, "/path/to/rasters/")
}
}
