% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rasters_from_library.R
\name{load_rasters_from_library}
\alias{load_rasters_from_library}
\title{Load Rasters From Library}
\usage{
load_rasters_from_library(start, end)
}
\arguments{
\item{start}{The start year for which raster files will be loaded.}

\item{end}{The end year for which raster files will be loaded.}
}
\value{
A list of raster files loaded from the specified library path for each year in the range from start to end.
}
\description{
This function loads raster files from a specified library path for a given range of years.
The raster files are expected to be named in the format "map-year.tif".
}
\examples{
\dontrun{
load_rasters_from_library(2000, 2005)
}
}
