% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_polygon.R
\name{make_polygon}
\alias{make_polygon}
\title{Create a Geographical Polygon with Specified Size and Shape}
\usage{
make_polygon(
  lat = NULL,
  lon = NULL,
  size = NULL,
  shape = c("circle", "square", "hexagon")
)
}
\arguments{
\item{lat}{A numeric latitude value of the center of the polygon.}

\item{lon}{A numeric longitude value of the center of the polygon.}

\item{size}{A numeric polygon size in kilometers.}

\item{shape}{A character string specifying the shape of the polygon, either "circle", "square", or "hexagon".}
}
\value{
The created polygon
}
\description{
Create a Geographical Polygon with Specified Size and Shape
}
\examples{
\dontrun{
make_polygon(lat = 40.7128, lon = 74.0060, size = 10, shape = "circle")
}
}
