% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_to_biomastats.R
\name{post_to_biomastats}
\alias{post_to_biomastats}
\title{Send Fragment Data to Biomastats API}
\usage{
post_to_biomastats(fragment, year, url = NULL)
}
\arguments{
\item{fragment}{A numeric or character value representing the fragment ID.}

\item{year}{A numeric value representing the year associated with the fragment.}

\item{url}{An optional character string specifying the URL for the API. Defaults to the Biomastats API URL.}
}
\value{
The content of the API response in JSON format as a character string, or an error message if the request fails.
}
\description{
This function sends a POST request to the Biomastats API with the specified fragment and year. 
It handles different URL options and returns the response content in JSON format.
}
\examples{
\dontrun{
post_to_biomastats(fragment = 1, year = 2024)
post_to_biomastats(fragment = 2, year = 2023, url = "https://custom-api.com/api/link")
}
}
