% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_coords.R
\name{square_coords}
\alias{square_coords}
\title{Square Coordinates}
\usage{
square_coords(lat = lat, lon = lon, adj_size = adj_size)
}
\arguments{
\item{lat}{Latitude of the center of the square. Default is 'lat'.}

\item{lon}{Longitude of the center of the square. Default is 'lon'.}

\item{adj_size}{Adjustment size for the square. This will affect the size of the square. Default is 'adj_size'.}
}
\value{
A Polygon object representing the square.
}
\description{
This function generates the coordinates for a square given a center (latitude and longitude) and an adjustment size.
The square's vertices are calculated based on the geometric properties of a square.
}
\examples{
\dontrun{
square_coords(lat = 40.7128, lon = -74.0060, adj_size = 1)
}
}
