% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_maps.R
\name{download_maps}
\alias{download_maps}
\title{Download Maps}
\usage{
download_maps(num_year, fragment_id, file_name)
}
\arguments{
\item{num_year}{Year number of the map to be downloaded.}

\item{fragment_id}{ID of the map fragment to be downloaded.}
}
\value{
No return value, called for side effects.
}
\description{
This function downloads maps using a list of links.
}
\examples{
\dontrun{
  download_maps(num_year = 2022, fragment_id = 1, file_name)
}
}
