% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area.R
\name{get_area}
\alias{get_area}
\title{Calculate Land Classes Area}
\usage{
get_area(data = NULL)
}
\arguments{
\item{data}{A list containing two elements: 'time_range': a vector of two elements specifying the start and end years, and 'raster': a list of raster objects for each year in the time range.}
}
\value{
A list with four elements: 'Years':the vector of years, 'Occupied area': a list with the calculated areas for each land class and year, 'aggregate_data': a data frame with the calculated areas for each land class and year, 'time_series': a ggplot2 object with the time series of areas.
}
\description{
This function calculates the area of land classes over a specified time range.
}
\examples{
\dontrun{
data <- ... to specify ....
result <- get_area(data)
}
}
