% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rasters.R
\name{load_rasters}
\alias{load_rasters}
\title{Load Rasters from Shapefile and Raster Data}
\usage{
load_rasters(
  shape_path = NULL,
  start = 1985,
  end = 2020,
  data_from = c("download", "library", "folder", "example"),
  folder_path = NULL
)
}
\arguments{
\item{shape_path}{A character string specifying the path to the shapefile.}

\item{start}{A numeric value indicating the start year for raster data.}

\item{end}{A numeric value indicating the end year for raster data.}

\item{data_from}{A character vector specifying the source of raster data. Can be "download", "library", or "folder".}

\item{folder_path}{A character string specifying the path to the folder containing raster files. Required if data_from = "folder".}
}
\value{
A list containing the shapefile, time range, and masked raster data.
}
\description{
This function loads raster data from either an online source, a local folder, or a library.
It then crops and masks the raster data using a provided shapefile.
}
\examples{
\dontrun{
load_rasters("path/to/shapefile.shp", start = 1985, end = 2020, data_from = "library")
}
}
